Don't scale points, use 100x100 viewBox
Since the SVG is only being used to draw lines and paths which have
`vector-effect: non-scaling-stroke;` set we don't need to set a viewBox
that's anything like the real aspect ratio of the image.

This means we can scale the data we're showing to a nice simple 0-100
range. In the case of the stub data it's in that range anyway, so no
need for the scaling code for now - it would be needed in the real world
though.